Address review: consolidate builder, add fetch_tier to live snapshots
- Merge _build_betfair_entry/_build_bookmaker_entry into single function
  with optional betfair_matched kwarg (#1)
- Compute fetch_tier and hours_until_commence on live snapshots so they
  participate in training pipeline filtering (#6)
- Extract hours_to_tier to oddsportal_common.py (shared by historical
  ingest and live collection), remove duplicate from ingest script
- Use extend instead of append loop (#nit)
- Add test_whole_number for fractional_to_decimal (#4)
